+++ index.html
@@ -8,7 +8,7 @@
     <title>The Endless Expedition</title>
 </head>
 <body>
-<!--comment-->
+
 <div id="game"></div>
 
 <script type="module" src="main.ts"></script>
