+++ src/classes/resourcesDescription.ts
@@ -1,0 +1,14 @@
+interface ResourcesDescription {
+    [key: string]: string
+}
+export const resourcesDescription: ResourcesDescription = {
+    grass: 'Обычная трава. Ничего необычного.',
+    stick: 'Жаль, не похожа на автомат, которым ты играл в детстве.',
+    commonFlower: 'Охапка полевых цветов. С ними ты выглядишь как жених.',
+    goldenFlower: `Второе название - Золотой цветок. Очень редкий. Назван так из свойства переливаться
+                 характерным свечением на Солнце. Лепестки Златограйника обладают хорошими лечебными свойствами.`,
+    littleStone: `Камушек. Используется в большом количестве при игре в "Камушки". И нет, они не могут делать того, о чем
+                ты подумал. Произнеси сто раз, что бы потерять смысл слова.`,
+    distilledWater: `Мама говорила, что дистилированную воду пить опасно. Сейчас посмотрм, правда это или нет.`,
+    dryFood: 'На вкус как копыта коня вперемешку с его гривой.'
+}
