+++ src/classes/windowsVisibility.ts
@@ -1,0 +1,9 @@
+export interface WindowsVisibility {
+    mainWindow: boolean
+    activeMainWindow: string
+}
+
+export const windowsVisibilityObj: WindowsVisibility = {
+    mainWindow: true,
+    activeMainWindow: 'Inventory'
+}
