+++ src/components/Subwindows/Inventory.vue
@@ -1,0 +1,90 @@
+<script setup lang="ts">
+import {storeToRefs} from "pinia";
+import {computed, ref, toRaw, watch} from "vue";
+
+import {useGameStore} from "@/store/useGameStore.ts";
+import type {InventoryItemsTypes, InventoryResourceSimple} from "@/classes/player.ts";
+
+import {resourcesDescription} from "@/classes/resourcesDescription.ts";
+
+const gameStore = useGameStore()
+const {player, windowsVisibility} = storeToRefs(gameStore)
+const placeholderResource: InventoryResourceSimple = {
+  name: 'placeholder', engName: 'placeholder', count: 0, type: 'placeholder'
+}
+const activeResource = ref<InventoryItemsTypes>(placeholderResource)
+const activeResourceDescription = ref<string>('')
+
+const setActiveResource = (newResource: InventoryResourceSimple) => {
+  activeResource.value = toRaw(newResource)
+}
+const isActiveResourcePlaceholder = computed<boolean>(() => {
+  return activeResource.value.type === 'placeholder'
+})
+const isActiveResourceSimple = computed<boolean>(() => {
+  return activeResource.value.type === 'resource' || activeResource.value.type === 'food'
+})
+
+watch(() => activeResource.value, (newActiveResource) => {
+  if (newActiveResource && resourcesDescription[newActiveResource.engName]) {
+    activeResourceDescription.value = resourcesDescription[newActiveResource.engName]
+  } else {
+    activeResourceDescription.value = 'Описания еще нет'
+  }
+})
+</script>
+
+<template>
+  <div class="subWindow__header">Инвентарь</div>
+
+  <div class="resource__subWindow main__texture"
+       :class="{_show: windowsVisibility.mainWindow && player.inventory.playerInventory.length !== 0}">
+
+    <template v-if="isActiveResourcePlaceholder">
+      <p class="_little">Нажмите на ресурс, что бы увидеть способы взаимодействия с ним!</p>
+    </template>
+
+    <template v-else-if="isActiveResourceSimple">
+      <p class="_little">{{activeResourceDescription}}</p>
+    </template>
+  </div>
+
+  <p v-if="player.inventory.playerInventory.length === 0">Пуста! Даже повешенной мышки нет(</p>
+
+  <div v-else class="container _flex">
+    <ul class="inventory">
+      <li class="resource"
+          v-for="resourceUnit of player.inventory.playerInventory"
+          @click="setActiveResource(resourceUnit)" :key="resourceUnit.engName">
+        {{resourceUnit.name}}: {{resourceUnit.count}}
+      </li>
+    </ul>
+  </div>
+
+</template>
+
+<style>
+.resource {
+  cursor: pointer;
+}
+.inventory {
+  flex-direction: row;
+  display: flex;
+  gap: var(--gap);
+}
+.resource__subWindow {
+  margin-top: auto;
+  position: absolute;
+  bottom: 100%;
+  left: calc(var(--gap) * 2);
+  right: 0;
+  padding: var(--gap);
+  border-radius: var(--radius) 0 0 0;
+  border-right: none;
+  transition-duration: var(--transition);
+  transform: translateX(100%);
+}
+.resource__subWindow._show {
+  transform: translateX(0);
+}
+</style>
