+++ src/components/Subwindows/Recipes.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts"></script>
+
+<template>
+  рецепты
+</template>
+
+<style></style>
