+++ src/composables/useGetRandomByRange.ts
@@ -1,0 +1,5 @@
+export const useGetRandomByRange = (range: [number, number]): number => {
+    const [min, max] = range
+    
+    return Math.round(Math.random() * (max - min) + min);
+}
