+++ srcOld/components/LocationMenu/Info.vue
