+++ srcOld/components/Map/Location.vue
@@ -1,17 +1,14 @@
-<script setup lang="ts">
-import {useGameStore} from "@/store/useGameStore.ts";
-import {storeToRefs} from "pinia";
-import {computed} from "vue";
-import type { Location } from "@/classes/allLocations.ts";
-import type {InventoryResourceSimple} from "@/classes/player.ts";
+<script setup>
+/**
+ * Компонент отдельной локации */
 
-const gameStore = useGameStore()
-const {player, notifications} = storeToRefs(gameStore)
+import {computed, inject} from "vue";
 
-const definedProps = defineProps<{
-  location: Location
-}>()
-
+const player = inject('player')
+const definedProps = defineProps({
+  location: Object
+})
+const definedEmits = defineEmits(['openSubWindow'])
 const locationStyle = computed(() => {
   return {
     top: `${definedProps.location.coords[0]}px`,
@@ -24,40 +21,27 @@ const hoverEffectSrc = computed(() => {
   return `/src/assets/images/hoverLocations/${definedProps.location.engName}HoverEffect.png`
 })
 
-const farmResource = () => {
-  if (!gameStore.currentLocationObj) return
-
-  const currentLocationObj = gameStore.currentLocationObj
-  const cockedResource: InventoryResourceSimple | boolean = player.value.inventory.farmResource(currentLocationObj)
-
-  if (cockedResource) notifications.value.showNotification(cockedResource, 'farmResource')
-  if (!cockedResource) notifications.value.showNotification(cockedResource, 'notFarmResource')
-}
 </script>
+
+<!-- todo пока не делал анимацию фарма ресурсов. Думаю сделать так, что бы всплывала иконка ресурса и количество -->
+<!-- todo сделать для submenu локации свой компонент и выводить его через сложный список -->
 
 <template>
   <div class="location__container"
-       :class="{_active: location.isCurrent}"
-       :style="locationStyle"
-       @click="player.changeCurrentLocation(location.engName)"
-  >
-    <picture class="location__hover"
-             v-show="!location.isCurrent">
-      <img :src="hoverEffectSrc" alt="location-hover">
-    </picture>
+       :style="locationStyle" :class="{_active: location.isCurrent}" @click="player.changeLocation(location.engName)">
+    <picture v-show="!location.isCurrent" class="location__hover"><img :src="hoverEffectSrc" alt="hoverLocation"></picture>
 
-    <div class="location"
-         :class="{_hidden: !location.isCurrent}">
+    <div class="location" :class="{_hidden: !location.isCurrent}">
       <p class="location__here main__texture">Вы <br> тут</p>
 
       <ul class="location__submenu">
-        <template v-for="{id, title} of location.subMenu" :key="id">
-          <li v-if="id !== 'resources'"
-              class="locationMenu__unit main__texture">
-            {{title}}
+        <template v-for="menuUnit of location.submenu" :key="menuUnit.id">
+          <li class="locationMenu__unit main__texture"
+              v-if="menuUnit.id !== 'resources'" @click="$emit('openSubWindow', menuUnit.id)">
+            {{menuUnit.title}}
           </li>
-          <li v-else class="locationMenu__unit main__texture" @click="farmResource">
-            {{title}}
+          <li class="locationMenu__unit main__texture" v-else @click="player.farmResource">
+            {{menuUnit.title}}
           </li>
         </template>
       </ul>
@@ -89,6 +73,10 @@ const farmResource = () => {
   opacity: 0;
   transition-duration: var(--transition);
   display: flex;
+}
+.location__hover img {
+  width: 100%;
+  height: 100%;
 }
 .location__container:hover .location__hover {
   opacity: 1;
